[PLFM-9516] Address PR review feedback
- Add TransactionTemplate to SearchIndexStatusDaoImpl (tables DB pattern)
- Create SearchIndexLifecycleManager with handleCreate/handleUpdate/handleDelete
- Separate CREATE (full rebuild) vs UPDATE (no-op if exists) behavior
- State-driven delete concurrency: CREATING blocks delete with recoverable exception
- Fix zombie DELETING state: retry instead of no-op
- Set FAILED before best-effort AOSS cleanup (status table is source of truth)
- Remove admin user, use change message userId
- Slim worker to thin delegation (~80 lines)
- Add recoverable-exception logging at worker level
- Move query endpoints from SearchQueryController to SearchManagementController
- Add field limits documentation table to controller Javadoc
