+++ services/repository/src/main/java/org/sagebionetworks/repo/web/controller/SearchManagementController.java
@@ -4,6 +4,9 @@ import static org.sagebionetworks.repo.model.oauth.OAuthScope.modify;
 import static org.sagebionetworks.repo.model.oauth.OAuthScope.view;
 
 import org.sagebionetworks.repo.model.AuthorizationConstants;
+import org.sagebionetworks.repo.model.asynch.AsyncJobId;
+import org.sagebionetworks.repo.model.asynch.AsynchronousJobStatus;
+import org.sagebionetworks.repo.model.search.SearchQueryResults;
 import org.sagebionetworks.repo.model.search.table.BindSearchConfigToEntityRequest;
 import org.sagebionetworks.repo.model.search.table.ColumnAnalyzerOverride;
 import org.sagebionetworks.repo.model.search.table.ListColumnAnalyzerOverridesRequest;
@@ -16,11 +19,13 @@ import org.sagebionetworks.repo.model.search.table.ListTextAnalyzersRequest;
 import org.sagebionetworks.repo.model.search.table.ListTextAnalyzersResponse;
 import org.sagebionetworks.repo.model.search.table.SearchConfigBinding;
 import org.sagebionetworks.repo.model.search.table.SearchConfiguration;
+import org.sagebionetworks.repo.model.search.table.SearchIndexQuery;
 import org.sagebionetworks.repo.model.search.table.SearchIndexStatus;
 import org.sagebionetworks.repo.model.search.table.SynonymSet;
 import org.sagebionetworks.repo.model.search.table.TextAnalyzer;
 import org.sagebionetworks.repo.service.search.ColumnAnalyzerOverrideService;
 import org.sagebionetworks.repo.service.search.SearchConfigurationService;
+import org.sagebionetworks.repo.service.search.SearchIndexQueryService;
 import org.sagebionetworks.repo.service.search.SearchIndexStatusService;
 import org.sagebionetworks.repo.service.search.SynonymSetService;
 import org.sagebionetworks.repo.service.search.TextAnalyzerService;
@@ -171,6 +176,98 @@ import org.springframework.web.bind.annotation.ResponseStatus;
  * <code>organizationName</code>. The organization and name cannot be changed after creation.
  * Resource names must start with a letter and contain only letters, digits, and underscores.
  * </p>
+ *
+ * <h6>Search Queries</h6>
+ * <p>
+ * Query a <a href="${org.sagebionetworks.repo.model.search.table.SearchIndex}">SearchIndex</a>
+ * using the async job pattern. Submit a
+ * <a href="${org.sagebionetworks.repo.model.search.table.SearchIndexQuery}">SearchIndexQuery</a>
+ * to start a job, then poll for
+ * <a href="${org.sagebionetworks.repo.model.search.SearchQueryResults}">SearchQueryResults</a>.
+ * A synchronous autocomplete endpoint is also available for typeahead patterns.
+ * </p>
+ * <ul>
+ * <li><a href="${POST.search.query.async.start}">POST /search/query/async/start</a> &mdash; Start async query</li>
+ * <li><a href="${GET.search.query.async.get.asyncToken}">GET /search/query/async/get/{asyncToken}</a> &mdash; Poll for results</li>
+ * <li><a href="${POST.search.autocomplete}">POST /search/autocomplete</a> &mdash; Synchronous autocomplete (max 8 results)</li>
+ * </ul>
+ *
+ * <h6>Search Index Field Limits</h6>
+ * <p>
+ * When a search index is built, each column from the defining SQL query is mapped to an
+ * OpenSearch field type. Text and keyword fields have an <code>ignore_above</code> limit on
+ * their keyword sub-field &mdash; values longer than this limit are <b>not indexed</b> for
+ * exact-match or sorting, but remain stored in the source document. Numeric, boolean, and
+ * JSON fields have no such limit.
+ * </p>
+ * <b>Search Index Field Mapping and Limits</b>
+ * <table border="1">
+ * <tr>
+ * <th>Synapse Column Type</th>
+ * <th>OpenSearch Field Type</th>
+ * <th>Keyword ignore_above</th>
+ * <th>Default Analyzer</th>
+ * </tr>
+ * <tr>
+ * <td>STRING, STRING_LIST</td>
+ * <td>text + keyword sub-field</td>
+ * <td>1,000 characters</td>
+ * <td>SCIENTIFIC</td>
+ * </tr>
+ * <tr>
+ * <td>MEDIUMTEXT</td>
+ * <td>text + keyword sub-field</td>
+ * <td>2,000 characters</td>
+ * <td>SCIENTIFIC</td>
+ * </tr>
+ * <tr>
+ * <td>LARGETEXT</td>
+ * <td>text + keyword sub-field</td>
+ * <td>8,192 characters</td>
+ * <td>SCIENTIFIC</td>
+ * </tr>
+ * <tr>
+ * <td>LINK</td>
+ * <td>text + keyword sub-field (or keyword + searchable, depending on analyzer)</td>
+ * <td>1,000 characters</td>
+ * <td>KEYWORD</td>
+ * </tr>
+ * <tr>
+ * <td>ENTITYID, USERID, ENTITYID_LIST, USERID_LIST</td>
+ * <td>keyword</td>
+ * <td>256 characters</td>
+ * <td>KEYWORD</td>
+ * </tr>
+ * <tr>
+ * <td>INTEGER, DATE, INTEGER_LIST, DATE_LIST, FILEHANDLEID, SUBMISSIONID, EVALUATIONID</td>
+ * <td>long</td>
+ * <td>N/A</td>
+ * <td>KEYWORD</td>
+ * </tr>
+ * <tr>
+ * <td>DOUBLE</td>
+ * <td>double</td>
+ * <td>N/A</td>
+ * <td>KEYWORD</td>
+ * </tr>
+ * <tr>
+ * <td>BOOLEAN, BOOLEAN_LIST</td>
+ * <td>boolean</td>
+ * <td>N/A</td>
+ * <td>KEYWORD</td>
+ * </tr>
+ * <tr>
+ * <td>JSON</td>
+ * <td>object (dynamic mapping)</td>
+ * <td>N/A</td>
+ * <td>STANDARD</td>
+ * </tr>
+ * </table>
+ * <p>
+ * <b>Query Limits:</b> Results per page default to 25 with a maximum of 100.
+ * Autocomplete results are capped at 8. The maximum number of rows that can be indexed
+ * in a single search index is 500,000.
+ * </p>
  */
 @ControllerInfo(displayName = "Search Management Services", path = "repo/v1")
 @Controller
@@ -191,6 +288,9 @@ public class SearchManagementController {
 
 	@Autowired
 	private SearchIndexStatusService searchIndexStatusService;
+
+	@Autowired
+	private SearchIndexQueryService searchIndexQueryService;
 
 	// ==================== Text Analyzers ====================
 
@@ -694,5 +794,75 @@ public class SearchManagementController {
 			@RequestParam(value = AuthorizationConstants.USER_ID_PARAM) Long userId,
 			@PathVariable String id) {
 		return searchIndexStatusService.getSearchIndexStatus(userId, id);
+	}
+
+	// ==================== Search Queries ====================
+
+	/**
+	 * Start an asynchronous search query job against a
+	 * <a href="${org.sagebionetworks.repo.model.search.table.SearchIndex}">SearchIndex</a>.
+	 * <p>
+	 * Use <a href="${GET.search.query.async.get.asyncToken}">GET /search/query/async/get/{asyncToken}</a>
+	 * to poll for results.
+	 * </p>
+	 *
+	 * @param userId The ID of the authenticated user.
+	 * @param request The search query request including the searchIndexId and query parameters.
+	 * @return An async job token to poll for results.
+	 */
+	@RequiredScope({ view })
+	@ResponseStatus(HttpStatus.CREATED)
+	@RequestMapping(value = UrlHelpers.SEARCH_QUERY_ASYNC_START, method = RequestMethod.POST)
+	public @ResponseBody AsyncJobId startQuery(
+			@RequestParam(value = AuthorizationConstants.USER_ID_PARAM) Long userId,
+			@RequestBody SearchIndexQuery request) {
+		AsynchronousJobStatus job = searchIndexQueryService.startSearchQuery(userId, request);
+		AsyncJobId asyncJobId = new AsyncJobId();
+		asyncJobId.setToken(job.getJobId());
+		return asyncJobId;
+	}
+
+	/**
+	 * Get the results of a previously started asynchronous search query.
+	 * <p>
+	 * Note: When the result is not ready yet, this method will return a status
+	 * code of 202 (ACCEPTED) and the response body will be a
+	 * <a href="${org.sagebionetworks.repo.model.asynch.AsynchronousJobStatus}">AsynchronousJobStatus</a> object.
+	 * </p>
+	 *
+	 * @param userId The ID of the authenticated user.
+	 * @param asyncToken The token returned by the start query endpoint.
+	 * @return The search query results.
+	 * @throws Throwable
+	 */
+	@RequiredScope({ view })
+	@ResponseStatus(HttpStatus.OK)
+	@RequestMapping(value = UrlHelpers.SEARCH_QUERY_ASYNC_GET, method = RequestMethod.GET)
+	public @ResponseBody SearchQueryResults getQueryResults(
+			@RequestParam(value = AuthorizationConstants.USER_ID_PARAM) Long userId,
+			@PathVariable String asyncToken) throws Throwable {
+		AsynchronousJobStatus jobStatus = searchIndexQueryService.getSearchQueryResults(userId, asyncToken);
+		return (SearchQueryResults) jobStatus.getResponseBody();
+	}
+
+	/**
+	 * Perform a synchronous autocomplete search query against a
+	 * <a href="${org.sagebionetworks.repo.model.search.table.SearchIndex}">SearchIndex</a>.
+	 * <p>
+	 * This endpoint uses a PREFIX query type and caps results at 8. It is intended for
+	 * typeahead / autocomplete UI patterns where low latency and small result sets are preferred.
+	 * </p>
+	 *
+	 * @param userId The ID of the authenticated user.
+	 * @param request The search query request including the searchIndexId.
+	 * @return The autocomplete results.
+	 */
+	@RequiredScope({ view })
+	@ResponseStatus(HttpStatus.OK)
+	@RequestMapping(value = UrlHelpers.SEARCH_AUTOCOMPLETE, method = RequestMethod.POST)
+	public @ResponseBody SearchQueryResults autocomplete(
+			@RequestParam(value = AuthorizationConstants.USER_ID_PARAM) Long userId,
+			@RequestBody SearchIndexQuery request) {
+		return searchIndexQueryService.autocomplete(userId, request);
 	}
 }
